Update program structure
I failed to plan very far ahead when beginning work on this, and this commit
remedies that. Having a 'Row' struct to store all the details of a row makes
if all you're working with is rows, but it's got some problems:

- The 'Row' struct previously (before this commit) stores hints and number of
  hints (array) and the row's contents (ull mask); storing the content is
  unnecessary since it's only used to calculate the hints
- If we're storing hints in structs, we'll need a struct for the columns, too,
  but it'd only store hints. Thus, I changed the 'Row' struct into a 'Hint'
  struct to be used for both Rows and Columns. To simplify the hint finding
  code, 'Hint' structs are linked lists

I also added stubs for upcoming implementation and a 'Game' struct which stores
the game's (global) state; both I did to further clarify the new structure.