Add storing of full level contents for calculation of column hints
To calculate the hints for a column, the character in that column in each row
needed, and cannot (without writing a very complicated function) be passed
incrementally. Since we read the level line-by-line (with each line
representing a row), we cannot at this time calculate the column hints. This
means that we must store the level data (rereading it would be inefficient)
for later use in Hint creation.