Complete board state printing code.
A macro to return a character representing the state of a cell as
recorded by the `filledmasks' and `rejectmasks'  members of `game'
was added. It was then used to replace the expression we previously
used, which returned an integer representing the status of a cell in
the original level data. Now we can print the board state as it will
appear in the final game rather than just reformating the level text
file.