Lay foundation for more versatile board printing logic
The old board printing logic was extremely was very limited and used
excessively complex (to the point of inpenetrability) code. This new
method aims to simplify the code by having a buffer to store the board
state in a printable, always up-to-date fashion, avoiding the complex
squences of control characters that was formerly used.

Functions used in the old method have been marked as deprecated, but
not removed; we would like to keey the program in a working state as
we make changes and only remove functionality when we are replacing
with a just-as-good alternative