Bind keys and rework cell selection logic
Since cell marking references a 2D array of only cell status bits its
coordinate system is zeroed at the first such bit and scaled so that
adjacent cells are referenced by adjacent integers. This is different
from the moving logic, which works on a 2D array which includes not
only the cells, but the padding between them, the hits, and invisible
control characters wrapping each visible character. In order to use
game.pos[xy] for both, some translation is necessary. I placed the
translation in the `selectrow' and `selectcol' functions, but the
code I used is very messy and repetitive and I hope to clean it up
soon.

The reason I have not cleaned it up now is a) it's functional and b)
the particular abstraction I have in mind will necessitate changing
code whose function lies beyond the scope of this commit.